Fix GCC c++11 narrowing issue
The previous definition of v4sil was trying to cast to a 64bit type, not not exactly matching the expected underlying type (long long).

Under c++11, this fails to compile with a narrowing warning.
i.e. in gcc when using the flags "-std=c++11 -Wnarrowing"

By casting to a "long long" instead of "unsigned long long" above, we match the type of the underlying definition, and also the v2dil cast then work correctly to widen the 2 64bit type to the expected final 128bit type.